HooksTest: rename to `DispatchTest`
... as the remaining tests relate directly to the `dispatch()` method.

Including re-ordering the methods in the class to follow the order of the code in the method under test.
